Cipher cache should include encryption method in key ... otherwise two servers with different encryption method but same password will share the same cipher, and client will fail to authenticate.
Instead of using map[string]map[string]*ss.Cipher, just prepend password
with encryption method name and a delimiter "|" as cache key. Since there
are no encryption methods with "|" in their names (and unlikely any further
ones), it's safe to use this delimiter.

Another problem is that, no method provided ("") is equivalent to "table",
but will have two entries in cipher cache. But it's hard to know the default
encryption method from package ss, and "table" is rarely used, this won't
be a big problem.